+++ lib/app/context.dart
@@ -79,7 +79,6 @@ extension AppContext on BuildContext {
       clientRepository.updateSeen(state).then((_) async {
         seenRepository.sync();
         app.syncComplete();
-        // TODO this seems like a lot...
         await reloadCounts();
         await unread.reload();
         await latest.reload();